Hash map is now a linked list.
While a hash map offers better performance in theory a linked list is
superior for the ordering of the nodes. I could eventually figure out an
ordered hash map but at the end of the day a maximum of 6 to 8 nodes at
O(1) performance is not going to make the time spent writing that worth
it.