Changed function signature of sanitize function.
New parameter _initial_str is copied into _str memory which is a member
of player info struct. This only happens after it checks if it is
necessary.

New return code. Most of the time it will return 0. If it returns the
character code for 'X' then it means the X social media website has
provided metadata to playerctl. This is difficult to parse so the block
will display "Website 'X' is playing media".